requirements.txt
streamlit
openai
llama-index
llama-index-llms-openai
nltk
pypdf
weaviate-client
llama_index.readers.weaviate
llama-index-vector-stores-weaviate
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai